[IMP] hr_holidays_public: Leave only public holidays computation
Reduce to the minimum the footprint of this module, adding only public
holidays as leaves in a transparent way by other modules.
