+++ hr_holidays_public/models/__init__.py
@@ -1,6 +1,4 @@
-from . import hr_employee
 from . import hr_holidays
 from . import hr_holidays_status
 from . import hr_holidays_public
-from . import hr_holidays_public_line
 from . import resource
