+++ hr_holidays_public/models/hr_holidays_public.py
@@ -105,7 +105,6 @@ class HrHolidaysPublic(models.Model):
                                employee.address_id.state_id.id)]
         else:
             states_filter.append(('state_ids', '=', False))
-
         hhplo = self.env['hr.holidays.public.line']
         holidays_lines = hhplo.search(states_filter)
         return holidays_lines
@@ -129,3 +128,66 @@ class HrHolidaysPublic(models.Model):
             if hol_date.ids:
                 return True
         return False
+
+
+class HrHolidaysPublicLine(models.Model):
+    _name = 'hr.holidays.public.line'
+    _description = 'Public Holidays Lines'
+    _order = "date, name desc"
+
+    name = fields.Char(
+        'Name',
+        required=True,
+    )
+    date = fields.Date(
+        'Date',
+        required=True
+    )
+    year_id = fields.Many2one(
+        'hr.holidays.public',
+        'Calendar Year',
+        required=True,
+    )
+    variable_date = fields.Boolean('Date may change', oldname='variable',
+                                   default=True)
+    state_ids = fields.Many2many(
+        'res.country.state',
+        'hr_holiday_public_state_rel',
+        'line_id',
+        'state_id',
+        'Related States'
+    )
+
+    @api.multi
+    @api.constrains('date', 'state_ids')
+    def _check_date_state(self):
+        for line in self:
+            line._check_date_state_one()
+
+    def _check_date_state_one(self):
+        if fields.Date.from_string(self.date).year != self.year_id.year:
+            raise UserError(_(
+                'Dates of holidays should be the same year '
+                'as the calendar year they are being assigned to'
+            ))
+
+        if self.state_ids:
+            domain = [('date', '=', self.date),
+                      ('year_id', '=', self.year_id.id),
+                      ('state_ids', '!=', False),
+                      ('id', '!=', self.id)]
+            holidays = self.search(domain)
+
+            for holiday in holidays:
+
+                if self.state_ids & holiday.state_ids:
+                    raise UserError(_('You can\'t create duplicate public '
+                                      'holiday per date %s and one of the '
+                                      'country states.') % self.date)
+        domain = [('date', '=', self.date),
+                  ('year_id', '=', self.year_id.id),
+                  ('state_ids', '=', False)]
+        if self.search_count(domain) > 1:
+            raise UserError(_('You can\'t create duplicate public holiday '
+                              'per date %s.') % self.date)
+        return True
