+++ hr_holidays_public/readme/CONFIGURE.rst
@@ -1,5 +1,3 @@
 Go to *Leaves -> Configuration* and open a Leave Type
 
-* Check "Exclude Rest Days" to exclude resource calendar rest days
-* Check "Exclude Public Holidays" to exclude public holidays
-* Uncheck "Compute Full Days" to allow fractional days.
+* Check "Exclude Public Holidays" to exclude public holidays.
