+++ hr_holidays_public/readme/DESCRIPTION.rst
@@ -1,4 +1,4 @@
 This module is a technical module to handle public holidays.
-The calculation of each leave can exclude rest days or public holiday.
-The calculation of each leave includes the number of hours taken from 
-employee calendar attendances
+
+The calculation of each leave can exclude rest public holiday, depending on
+the leave type configuration.
