+++ hr_holidays_public/static/description/index.html
@@ -368,10 +368,9 @@ ul.auto-toc {
 !! changes will be overwritten.                   !!
 !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!! -->
 <p><a class="reference external" href="https://odoo-community.org/page/development-status"><img alt="Beta" src="https://img.shields.io/badge/maturity-Beta-yellow.png" /></a> <a class="reference external" href="http://www.gnu.org/licenses/agpl-3.0-standalone.html"><img alt="License: AGPL-3" src="https://img.shields.io/badge/licence-AGPL--3-blue.png" /></a> <a class="reference external" href="https://github.com/OCA/hr/tree/11.0/hr_holidays_public"><img alt="OCA/hr" src="https://img.shields.io/badge/github-OCA%2Fhr-lightgray.png?logo=github" /></a> <a class="reference external" href="https://translation.odoo-community.org/projects/hr-11-0/hr-11-0-hr_holidays_public"><img alt="Translate me on Weblate" src="https://img.shields.io/badge/weblate-Translate%20me-F47D42.png" /></a> <a class="reference external" href="https://runbot.odoo-community.org/runbot/116/11.0"><img alt="Try me on Runbot" src="https://img.shields.io/badge/runbot-Try%20me-875A7B.png" /></a></p>
-<p>This module is a technical module to handle public holidays.
-The calculation of each leave can exclude rest days or public holiday.
-The calculation of each leave includes the number of hours taken from
-employee calendar attendances</p>
+<p>This module is a technical module to handle public holidays.</p>
+<p>The calculation of each leave can exclude rest public holiday, depending on
+the leave type configuration.</p>
 <p><strong>Table of contents</strong></p>
 <div class="contents local topic" id="contents">
 <ul class="simple">
@@ -390,15 +389,29 @@ employee calendar attendances</p>
 <h1><a class="toc-backref" href="#id1">Configuration</a></h1>
 <p>Go to <em>Leaves -&gt; Configuration</em> and open a Leave Type</p>
 <ul class="simple">
-<li>Check “Exclude Rest Days” to exclude resource calendar rest days</li>
-<li>Check “Exclude Public Holidays” to exclude public holidays</li>
-<li>Uncheck “Compute Full Days” to allow fractional days.</li>
+<li>Check “Exclude Public Holidays” to exclude public holidays.</li>
 </ul>
 </div>
 <div class="section" id="usage">
 <h1><a class="toc-backref" href="#id2">Usage</a></h1>
-<p>Go to the menu <em>Leaves &gt; Public Holidays &gt; Public Holidays</em> and create your
-public holidays.</p>
+<p>For adding public holidays:</p>
+<ol class="arabic simple">
+<li>Go to the menu <em>Leaves &gt; Public Holidays &gt; Public Holidays</em>.</li>
+<li>Create your public holidays.</li>
+</ol>
+<p>For using public holidays on leaves:</p>
+<ol class="arabic simple">
+<li>Go to <em>Leaves &gt; Dashboard</em>.</li>
+<li>Select dragging on the calendar the days you want to be on leave, or go
+to the form view for selecting start and end dates.</li>
+<li>Select the proper “Leave Type” that has “Exclude Public Holidays” checked.</li>
+<li>If no leave type is yet specified, then default configuration is to exclude
+public holidays.</li>
+<li>The number of days will be computed excluding public holidays that match the
+selected employee, including global, country and state holidays.</li>
+<li>If no employee is yet selected, only global holidays will be taken into
+account.</li>
+</ol>
 </div>
 <div class="section" id="bug-tracker">
 <h1><a class="toc-backref" href="#id3">Bug Tracker</a></h1>
@@ -415,6 +428,7 @@ If you spotted it first, help us smashing it by providing a detailed and welcome
 <ul class="simple">
 <li>Michael Telahun Makonnen</li>
 <li>Tecnativa</li>
+<li>Fekete Mihai (Forest and Biomass Services Romania)</li>
 </ul>
 </div>
 <div class="section" id="contributors">
